conf: Templates for bblayers.conf and local.conf
Create templates for bblayers.conf and local.conf as per the yocto
documentation [1]. This replaces assembling the files from config fragments
in the ws_init.sh script.

The reason for keeping these confiuration files in favor of a
trustx-distro for setting machine specific configurations is that these
files set (among others) TRUSTME_CONTAINER_ARCH. This is used to
determine the container config via multiconfig. As multiconfig is parsed
before distro configs, this needs to be set very early in the parsing
process (e.g. in local.conf).

[1] https://docs.yoctoproject.org/singleindex.html#build-conf-bblayers-conf

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>