cml recipes: conditionally build from externalsrc
Let the recipe check, if the source repositoriy is checkout out at the
well-known path. If so, build from this directory using externalsrc.
Allows pinning the cml version in the repo manifest. If the repository
is missing, build upstream from git.

Many recipes in our layers depend on the cml repo being placed at the
well-known location (e.g. protobuf files trustx-signing). To obtain the
proto files independent whether the cml repo is there or not, install
the proto files to /usr/include/proto and obtain them from there.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>