recipes-trustx/*: Build cml-related recipes out-of-tree
We require cmld and service* to be built using dfferent variants of libc
for container and host builds (e.g. glibc and musl). Build in-source poisons
the source directory with artifacts compiled with another libc and thus
breaks the build if building glibc containers. Thus, we need to keep the
source directory clean and copy the sources to the build directory first.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>