recipes-initscripts/cml-boot: Add missing file for SSH server start
This file contains the commands to start the openSSH server in the cml
layer. If the feature is activated (in dev builds) its contents are
copied into the init script.

In previous commits regarding the openSSH integration into the cml layer
this file was accidentally not commited resulting in the SSH server not
starting at boot. This is hereby fixed.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>