recipes-initscripts/cml-boot: Create '/data/logs' before start of tpm2d
`tmp2d` looks to open log files in `/data/logs` When the logs folder is
not present, the `tpm2d`. segfaults and crashes a freshly installed OS.
This only happens when a TPM2.0 is enabled and present.  Moving the
`mkdir` of `/data/logs` up ahead of where the `tpm2d -n` is launched to
allow for a successful boot.

Signed-off-by: Ben Corrado <ben@nerdnode.io>
[ Michael Weiß: formatting and edit heading of commit message ]
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>