recipes-initscripts: cml-boot-script: FDE re-encrypt with AES256
If legacy FDE with AES128 is detected, also do the otf-encryption
similar to initial plain to AES256 encryption. We had a bug in
common/cryptfs which accidentally truncated the AES key to 256 bit.
This resulted in the use of AES128 since this is used as input for
XTS-mode. We now try to setup encryption as before using AES256. If
that does not succeed, we also try the legacy half-sized key for
AES128. If the mount then succeeds, we do the re-encryption.
The nvindex in the TPM already has the correct length as this was
always created with 512 bit random data.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>