recipes-initscript/cml-boot: Start migration to fragment based build
The cml-boot script contains increasingly more lines where code snippets
are patched in conditionally. This is particularly complicated if this
comes from other layers as still the placeholder value needs to be
inserted in meta-gyroidos. This commit starts migrating the compile
process to a fragment based flow similar to traditional System-V init
scripts. Init script fragments are added to the build directory, either
copied from source or put there by a yocto task. The init script is then
assembled by cat'ing the fragments in alphabetical order.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>