recipes-initscripts/cml-boot: Kill init if mount firmware/modules fails
In case of an error when mounting the modules/firmware image, the
initscirpt will continue starting the cmld and the c0. This can lead to
unintentionally acknowldging an update even if image mounting failed. We
kill init in case of failed mount since the image will hardly be usable
without modules or firmware.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>